requirements.txt
confluent-kafka 
faker 
pandas 
matplotlib 
seaborn
psycopg2-binary
sqlalchemy